+++ packages/backend-platform-test-stubs/src/index.ts
@@ -3,3 +3,4 @@ export * from './construct_container_stub.js';
 export * from './import_path_verifier_stub.js';
 export * from './stack_resolver_stub.js';
 export * from './ssm_environment_entries_generator_stub.js';
+export * from './resource_name_validator_stub.js';