+++ packages/integration-tests/src/constants.ts
@@ -1,6 +1,4 @@
 /**
- * Used to run `create amplfiy` at a specific tagged release
- * Remove @beta before GA to switch to running against @latest
- * https://github.com/aws-amplify/amplify-backend/issues/1013
+ * Used to run `create amplify` at a specific tagged release
  */
-export const amplifyAtTag = 'amplify@beta';
+export const amplifyAtTag = 'amplify@latest';