+++ packages/plugin-types/src/resource_name_validator.ts
@@ -1,0 +1,9 @@
+/**
+ * Implementors can validate what should be the correct resource name
+ */
+export type ResourceNameValidator = {
+  /**
+   * executes the validation
+   */
+  validate: (resourceName: string) => void;
+};